+++ netpyne/simFuncs.py
@@ -559,27 +559,27 @@ def cellByGid(gid):
 def readCmdLineArgs():
     import imp, __main__
 
-    print '\nReading command line arguments using syntax: python file.py [simConfig_path [netParams_path]]'
+    print '\nReading command line arguments using syntax: python file.py [simConfig=filepath] [netParams=filepath]'
 
-    argv = sys.argv
+    cfgPath = None
+    netParamsPath = None
+
     # read simConfig and netParams paths
-    if len(argv) > 2:  
-        cfgPath = argv[1]
-        netParamsPath = argv[2]
-        cfg = sim.loadSimCfg(cfgPath, setLoaded=False)
-        netParams = sim.loadNetParams(netParamsPath,  setLoaded=False)
-    # read simConfig path and use netParams.py
-    elif len(argv) > 1:  
-        cfgPath = argv[1]
-        cfg = sim.loadSimCfg(cfgPath, setLoaded=False)
-        __main__.cfg = cfg
-        netParamsModule = imp.load_source('netParams', 'netParams.py')
-        netParams = netParamsModule.netParams
-    else: 
-        # use simConfig.py and netParams.py
+    for arg in sys.argv:
+        if arg.startswith('simConfig='):  
+            cfgPath = arg.split('simConfig=')[1]
+            cfg = sim.loadSimCfg(cfgPath, setLoaded=False)
+            __main__.cfg = cfg
+        elif arg.startswith('netParams='):  
+            netParamsPath = arg.split('netParamsPath=')[1]
+            netParams = sim.loadNetParams(netParamsPath,  setLoaded=False)
+
+    if not cfgPath:
         cfgModule = imp.load_source('cfg', 'cfg.py')  
         cfg = cfgModule.cfg
         __main__.cfg = cfg
+
+    if not netParamsPath:
         netParamsModule = imp.load_source('netParams', 'netParams.py')
         netParams = netParamsModule.netParams